+++ enclave-manager/api/typescript/src/kurtosis_enclave_manager_api_connect.ts
@@ -3,10 +3,10 @@
 /* eslint-disable */
 // @ts-nocheck
 
-import { GetListFilesArtifactNamesAndUuidsRequest, GetServicesRequest, GetStarlarkRunRequest, HealthCheckRequest, HealthCheckResponse, InspectFilesArtifactContentsRequest, RunStarlarkPackageRequest } from "./kurtosis_enclave_manager_api_pb.js";
+import { DownloadFilesArtifactRequest, GetListFilesArtifactNamesAndUuidsRequest, GetServicesRequest, GetStarlarkRunRequest, HealthCheckRequest, HealthCheckResponse, InspectFilesArtifactContentsRequest, RunStarlarkPackageRequest } from "./kurtosis_enclave_manager_api_pb.js";
 import { Empty, MethodKind } from "@bufbuild/protobuf";
 import { CreateEnclaveArgs, CreateEnclaveResponse, DestroyEnclaveArgs, GetEnclavesResponse, GetServiceLogsArgs, GetServiceLogsResponse } from "./engine_service_pb.js";
-import { GetServicesResponse, GetStarlarkRunResponse, InspectFilesArtifactContentsResponse, ListFilesArtifactNamesAndUuidsResponse, StarlarkRunResponseLine } from "./api_container_service_pb.js";
+import { GetServicesResponse, GetStarlarkRunResponse, InspectFilesArtifactContentsResponse, ListFilesArtifactNamesAndUuidsResponse, StarlarkRunResponseLine, StreamedDataChunk } from "./api_container_service_pb.js";
 
 /**
  * @generated from service kurtosis_enclave_manager.KurtosisEnclaveManagerServer
@@ -85,6 +85,15 @@ export const KurtosisEnclaveManagerServer = {
       I: InspectFilesArtifactContentsRequest,
       O: InspectFilesArtifactContentsResponse,
       kind: MethodKind.Unary,
+    },
+    /**
+     * @generated from rpc kurtosis_enclave_manager.KurtosisEnclaveManagerServer.DownloadFilesArtifact
+     */
+    downloadFilesArtifact: {
+      name: "DownloadFilesArtifact",
+      I: DownloadFilesArtifactRequest,
+      O: StreamedDataChunk,
+      kind: MethodKind.ServerStreaming,
     },
     /**
      * @generated from rpc kurtosis_enclave_manager.KurtosisEnclaveManagerServer.DestroyEnclave