+++ enclave-manager/web/src/components/theme/tabsTheme.ts
@@ -15,8 +15,11 @@ export const tabsTheme = defineMultiStyleConfig({
         fontStyle: "normal",
         fontWeight: "medium",
         fontSize: "lg",
-        color: "gray.200",
+        color: "gray.100",
         lineHeight: "28px",
+        _hover: {
+          bg: `gray.700`,
+        },
         _selected: {
           fontWeight: "semibold",
           color: `${props.colorScheme}.400`,